fix(runtime): handle stale endpoint artifacts during recovery
Treat non-socket and wrong-protocol artifacts as absent runtime endpoints while retaining errors for uncertain ownership. Exercise stale artifacts through SDK and maintenance reconciliation, and move reconciliation into its own module.
