+++ sdk/rust/lib/backend/local/sandbox/mod.rs
@@ -6,6 +6,7 @@
 //! methods; [`LocalBackend::create_sandbox`] is its entry point.
 
 mod create;
+mod reconciliation;
 
 use std::collections::{BTreeMap, HashMap, HashSet};
 use std::path::Path;
@@ -14,13 +15,9 @@ use std::time::Duration;
 
 use base64::{Engine as _, engine::general_purpose::URL_SAFE_NO_PAD};
 use futures::{StreamExt, future::BoxFuture, stream};
-use microsandbox_db::pool::DbPools;
 use microsandbox_db::{DbReadConnection, DbWriteConnection};
 use microsandbox_image::{Digest, GlobalCache};
 use microsandbox_protocol::message::MessageType;
-use microsandbox_runtime::transition::{
-    SandboxTransitionGuard, sandbox_runtime_endpoint_is_live, try_acquire_transition_guard,
-};
 use sea_orm::{
     ColumnTrait, Condition, EntityTrait, ExprTrait, QueryFilter, QueryOrder, QuerySelect,
     sea_query::Expr,
@@ -541,166 +538,6 @@ impl LocalBackend {
 //--------------------------------------------------------------------------------------------------
 
 impl LocalBackend {
-    /// Reconcile a Starting/Running/Draining row against the owning process's
-    /// liveness, marking it terminal when the runtime is gone.
-    async fn reconcile_sandbox_runtime_state(
-        &self,
-        pools: &DbPools,
-        sandbox: sandbox_entity::Model,
-    ) -> MicrosandboxResult<sandbox_entity::Model> {
-        let run_dir = self.config().run_dir();
-        let sandboxes_dir = self.config().sandboxes_dir();
-        Self::reconcile_sandbox_runtime_state_with_paths(
-            pools,
-            sandbox,
-            Some((&run_dir, &sandboxes_dir)),
-        )
-        .await
-    }
-
-    /// Reconcile runtime state with optional exact socket roots.
-    async fn reconcile_sandbox_runtime_state_with_paths(
-        pools: &DbPools,
-        sandbox: sandbox_entity::Model,
-        socket_roots: Option<(&Path, &Path)>,
-    ) -> MicrosandboxResult<sandbox_entity::Model> {
-        Self::reconcile_sandbox_runtime_state_with_transition(pools, sandbox, socket_roots, None)
-            .await
-    }
-
-    /// Reuse a start caller's transition ownership instead of trying to lock itself.
-    async fn reconcile_sandbox_runtime_state_with_transition(
-        pools: &DbPools,
-        sandbox: sandbox_entity::Model,
-        socket_roots: Option<(&Path, &Path)>,
-        held_transition: Option<&SandboxTransitionGuard>,
-    ) -> MicrosandboxResult<sandbox_entity::Model> {
-        if !matches!(
-            sandbox.status,
-            SandboxStatus::Starting | SandboxStatus::Running | SandboxStatus::Draining
-        ) {
-            return Ok(sandbox);
-        }
-
-        let run = Self::load_active_run(pools.read(), sandbox.id).await?;
-        if run
-            .as_ref()
-            .and_then(|run| run.pid)
-            .is_some_and(Self::pid_is_alive)
-        {
-            return Ok(sandbox);
-        }
-
-        // A launcher owns Starting before the child records its PID. Prove that
-        // handoff abandoned with the same lock used by existing launchers.
-        let transition_guard = if sandbox.status == SandboxStatus::Starting
-            && held_transition.is_none()
-            && let Some((run_dir, _)) = socket_roots
-        {
-            let Some(guard) = try_acquire_transition_guard(run_dir, &sandbox.name)? else {
-                return Ok(sandbox);
-            };
-            Some(guard)
-        } else {
-            None
-        };
-        let owns_transition = held_transition.is_some() || transition_guard.is_some();
-
-        // A dead-PID snapshot is not sufficient: another process may already
-        // have reconciled and restarted this name. Serialize on the runtime
-        // ownership lock, then re-read the exact row/run before unlinking.
-        let _guard = if let Some((run_dir, _)) = socket_roots {
-            let Some(guard) =
-                microsandbox_runtime::ipc::try_acquire_lifecycle_guard(run_dir, &sandbox.name)?
-            else {
-                return Ok(sandbox);
-            };
-            Some(guard)
-        } else {
-            None
-        };
-        let Some(sandbox) = sandbox_entity::Entity::find_by_id(sandbox.id)
-            .one(pools.read())
-            .await?
-        else {
-            return Err(crate::MicrosandboxError::SandboxNotFound(sandbox.name));
-        };
-        if !matches!(
-            sandbox.status,
-            SandboxStatus::Starting | SandboxStatus::Running | SandboxStatus::Draining
-        ) {
-            return Ok(sandbox);
-        }
-        let run = Self::load_active_run(pools.read(), sandbox.id).await?;
-
-        // Only an unowned start can be declared abandoned without a run row.
-        // Older runtimes may expose an endpoint without holding lifecycle locks.
-        let Some(run) = run else {
-            let abandoned_start = sandbox.status == SandboxStatus::Starting && owns_transition;
-            if abandoned_start
-                && let Some((run_dir, sandboxes_dir)) = socket_roots
-                && sandbox_runtime_endpoint_is_live(
-                    run_dir,
-                    &sandboxes_dir.join(&sandbox.name),
-                    &sandbox.name,
-                )?
-            {
-                return Ok(sandbox);
-            }
-            if sandbox.status == SandboxStatus::Draining || abandoned_start {
-                if let Some((run_dir, sandboxes_dir)) = socket_roots {
-                    crate::runtime::remove_sandbox_socket_artifacts_at(
-                        run_dir,
-                        sandboxes_dir,
-                        &sandbox.name,
-                    )?;
-                }
-                let (terminal_status, reason) = Self::stale_runtime_terminal_state(sandbox.status);
-                Self::mark_sandbox_runtime_stale(
-                    pools.write(),
-                    sandbox.id,
-                    None,
-                    terminal_status,
-                    reason,
-                )
-                .await?;
-
-                return sandbox_entity::Entity::find_by_id(sandbox.id)
-                    .one(pools.read())
-                    .await?
-                    .ok_or_else(|| crate::MicrosandboxError::SandboxNotFound(sandbox.name));
-            }
-
-            return Ok(sandbox);
-        };
-
-        if run.pid.is_some_and(Self::pid_is_alive) {
-            return Ok(sandbox);
-        }
-
-        if let Some((run_dir, sandboxes_dir)) = socket_roots {
-            crate::runtime::remove_sandbox_socket_artifacts_at(
-                run_dir,
-                sandboxes_dir,
-                &sandbox.name,
-            )?;
-        }
-        let (terminal_status, reason) = Self::stale_runtime_terminal_state(sandbox.status);
-        Self::mark_sandbox_runtime_stale(
-            pools.write(),
-            sandbox.id,
-            Some(run.id),
-            terminal_status,
-            reason,
-        )
-        .await?;
-
-        sandbox_entity::Entity::find_by_id(sandbox.id)
-            .one(pools.read())
-            .await?
-            .ok_or_else(|| crate::MicrosandboxError::SandboxNotFound(sandbox.name))
-    }
-
     /// Load the most recent active run record for a sandbox, if any.
     pub(crate) async fn load_active_run(
         db: &DbReadConnection,
@@ -1668,7 +1505,7 @@ mod tests {
     #[cfg(unix)]
     #[tokio::test]
     async fn abandoned_start_requires_launcher_and_runtime_ownership() {
-        for owner in ["none", "launcher", "runtime", "legacy", "caller"] {
+        for owner in ["none", "stale", "launcher", "runtime", "legacy", "caller"] {
             let temp = tempfile::Builder::new()
                 .prefix("msb-abandoned")
                 .tempdir_in("/tmp")
@@ -1704,6 +1541,12 @@ mod tests {
             } else {
                 None
             };
+            if owner == "stale" {
+                let path =
+                    microsandbox_runtime::ipc::sandbox_socket_paths(&run_dir, name).legacy_agent;
+                std::fs::create_dir_all(path.parent().unwrap()).unwrap();
+                std::fs::write(path, b"stale endpoint").unwrap();
+            }
             let legacy = if owner == "legacy" {
                 std::fs::create_dir_all(&run_dir).unwrap();
                 let path =
@@ -1730,7 +1573,7 @@ mod tests {
             )
             .await
             .unwrap();
-            let expected = if matches!(owner, "none" | "caller") {
+            let expected = if matches!(owner, "none" | "stale" | "caller") {
                 SandboxStatus::Crashed
             } else {
                 SandboxStatus::Starting
